+++ Examples/QuillDemo/QuillDemo/QuillDemoApp.swift
@@ -1,0 +1,12 @@
+import SwiftUI
+
+@main
+struct QuillDemoApp: App {
+    var body: some Scene {
+        WindowGroup {
+            NavigationStack {
+                ConfigView()
+            }
+        }
+    }
+}